Updating with additional example more typical of registration
The prior example was erroneous with the rotation happening at the
origin. A new test case there the fixed image has been translation
some, into a large image is a better test case. With this setup we
experiment with a composite transform there the first is a complete
solution and where it's only a partial, and observe the effects of
rotation about the center. This clearly demonstrates that the center
parameter for the transforms is in the fixed image's domain, not in a
intermediate range in the composite transform.