ENH: Four registration tests are added
Four registration tests are designed to show and evaluate the different
components of the registration process in SimpleITK.
A brief explanation about each process is put at the top of each page.
The results of the first two tests are evaluated with the results of
BRIANSFit that is run with the same parameters set. itkTestDriver is
used for comparison.
For the last two tests, a checkerboard filter is used before and after
registration.